backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
httpx==0.28.0
apscheduler==3.10.4
python-dotenv==1.0.1
nba_api==1.4.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.3
